# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ranke-db API
  version: 0.2.0
  description: |
    REST/HTTP binding for a single **ranke-db** stack: read the verifiable graph
    with a declarative query, and contribute signed claims to it.

    A server hosts exactly **one stack**, launched from a config file — there is
    no tenant or archive routing in the paths.

    ## Reads are queries

    The full read surface is `POST /query`: a tree-structured query (`select`
    generators, `where` filters, `output` shaping, `limit` bounds) as a JSON
    object. Cypher/GQL is **never** a client route — it is an internal execution
    engine the planner lowers a query to. A **cacheable GET subset** covers the by-id
    reads without a query body: `GET /{branch}/head`, `GET /{branch}/claim/{id}`,
    `GET /{branch}/claim/{id}/content`, and the privileged `GET /$universe/claim/{id}` /
    `GET /$universe/claim/{id}/content`. Content is addressed by the **claim** that
    holds it (not a raw hash), so whether the bytes are inline or a separate blob is
    hidden and the read is scoped to the claim's branch. Content also rides **inline**
    in query results via `output.content`; the content route fetches the bytes for a
    single claim — including the blob an `output.overflow: reference` stub names.

    ## Encodings and verifiability

    A query returns a **streaming sequence** of results, one item each, in the
    serialization named by `output.encoding`:

...

...

...

    ## Credentials and authorization

...
servers:
...
security:
  - {}
  - jwt: []
  - apikey: []
  - macaroon: []
...
paths:
...
  /{branch}/head:
    parameters:
      - $ref: "#/components/parameters/BranchName"
    get:
      operationId: getBranchHead
      summary: Current head id of a branch
      description: |
        Returns the branch's current head id — a moving target (it advances on every
        contribution). Cacheable **with revalidation** (weak `ETag`, `Cache-Control:
        no-cache`): a conditional request is cheap when the head has not moved. To
        inspect the head claim itself, fetch it via `GET /{branch}/claim/{id}`.
      tags: [read]
      responses:
        "200":
          description: The current head id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BranchHead" }
          headers:
            ETag:
              description: Weak validator for the current head.
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{branch}/claim/{id}:
    parameters:
      - $ref: "#/components/parameters/BranchName"
      - $ref: "#/components/parameters/ClaimId"
    get:
      operationId: getBranchClaim
      summary: Fetch a claim within a branch's closure
      description: |
        Returns claim `{id}` as its signed CBOR bytes, **only if it lies in branch
        `{name}`'s closure**. A claim that is superseded, contradicted, or otherwise
        outside the closure returns `404`, indistinguishable from one that does not
        exist. Immutably **cacheable** by id (strong `ETag`, `Cache-Control: public,
        immutable`): the id content-addresses the bytes, so they never change.
      tags: [read]
      responses:
        "200":
          description: The claim's signed CBOR bytes.
          content:
            application/cbor:
              schema:
                type: string
                format: binary
          headers:
            ETag:
              description: Strong validator — the claim id.
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /{branch}/claim/{id}/content:
    parameters:
      - $ref: "#/components/parameters/BranchName"
      - $ref: "#/components/parameters/ClaimId"
    get:
      operationId: getBranchClaimContent
      summary: Fetch the content of a claim within a branch's closure
      description: |
        Streams the content of claim `{id}` — **only if it lies in branch `{name}`'s
        closure** (same closure guarantee as the claim itself; out-of-closure or
        unknown → `404`). Content is addressed by the claim that holds it, not by a
        raw hash: the server resolves whether the bytes live inline in the claim or
        in a separate blob, so the client can't tell and doesn't need to — and the
        read is scoped to the claim's branch. Immutably **cacheable** by the claim id.
      tags: [read]
      responses:
        "200":
          description: The content bytes, streamed.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            ETag:
              description: Strong validator — the claim id.
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /$universe/claim/{id}:
    parameters:
      - $ref: "#/components/parameters/ClaimId"
    get:
      operationId: getUniverseClaim
      summary: Fetch a claim by id from the Universe (privileged)
      description: |
        Returns claim `{id}` as its signed CBOR bytes directly from the Universe,
        bypassing any branch table. This is a **privileged** head-id read, conferred
        only through the reserved `$universe` name (see core-access). Immutably
        cacheable by id.
      tags: [read]
      responses:
        "200":
          description: The claim's signed CBOR bytes.
          content:
            application/cbor:
              schema:
                type: string
                format: binary
          headers:
            ETag:
              description: Strong validator — the claim id.
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /$universe/claim/{id}/content:
    parameters:
      - $ref: "#/components/parameters/ClaimId"
    get:
      operationId: getUniverseClaimContent
      summary: Fetch the content of a claim by id from the Universe (privileged)
      description: |
        Streams the content of claim `{id}` directly from the Universe, bypassing
        any branch table — a **privileged** read conferred only through `$universe`.
        Content is addressed by the claim; whether the bytes are inline or a separate
        blob is hidden. Immutably cacheable by the claim id.
      tags: [read]
      responses:
        "200":
          description: The content bytes, streamed.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            ETag:
              description: Strong validator — the claim id.
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /system/verification:
    get:
      operationId: listVerifications
      summary: List verification runs
      description: |
        Verification runs across the whole stack, newest first. Each report is a
        **point-in-time record**: a layer repaired externally shows clean in a later
        run, so reports accumulate rather than overwrite — until explicitly removed
        with `DELETE /system/verification/{reportId}`.

        The API lives under `/system/` (rather than the paper's root `/verification…`)
        for two reasons: a root-level `/verification/{id}` would collide with
        `/{branch}/head` under net/http (both match `/verification/head`, neither more
        specific), and a run is a stack-wide operational resource — it roots at a
        closure named in the request body, not in the path.
      tags: [system]
      responses:
        "200":
          description: The verification reports.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationReportList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: startVerification
      summary: Start a verification run
      description: |
        Starts a run from a `VerificationConfig`: walk the closure rooted at
        `closure` (a branch name — resolved to its current head and **pinned** for
        the life of the run — or a head id directly), reading the named `layer`
        directly, re-checking each claim to the configured depth. It returns
        *findings*, not contents, so it never leaks across closures.

        A run may take a **very long time** (hours to days over a large closure at
        full-content depth), so it is always **asynchronous**: the call returns `202`
        immediately with the running report and a `Location` header pointing at the
        report resource. Poll it with `GET /system/verification/{reportId}`, pacing
        by the `Retry-After` hint; stop it with `DELETE`. Starting the same run twice
        yields two independent point-in-time reports (runs are not deduplicated).

        Verification is resource-heavy, so the stack caps the number of runs that may
        execute **concurrently** (configured; default 1). When that cap is already
        reached the call returns `429` — the server never stops a run to make room. To
        proceed, either wait (per `Retry-After`) or `GET /system/verification` to see
        what is running and free a slot deliberately: `cancel` a run to stop it while
        keeping its report, or `DELETE` it to remove it entirely.
      tags: [system]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerificationConfig" }
      responses:
        "202":
          description: The run started; the running report is returned.
          headers:
            Location:
              description: The report resource to poll — `/system/verification/{reportId}`.
              schema: { type: string }
            Retry-After:
              description: Suggested seconds to wait before the first poll.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationReport" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429":
          description: |
            The concurrent-run cap is reached. No run was started and none was
            cancelled; retry after a running one finishes or is stopped.
          headers:
            Retry-After:
              description: Suggested seconds to wait before retrying.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /system/verification/{reportId}:
    parameters:
      - name: reportId
        in: path
        required: true
        description: The verification report id.
        schema: { type: string }
    get:
      operationId: getVerification
      summary: Show a verification run
      description: |
        The report for a run. Poll until `status` leaves `running`; while it is still
        running the response carries a `Retry-After` hint and the progress counters
        (`claimsChecked`, `bytesRead`) advance between polls.
      tags: [system]
      responses:
        "200":
          description: The verification report.
          headers:
            Retry-After:
              description: While `status` is `running`, suggested seconds before the next poll.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteVerification
      summary: Delete a verification run
      description: |
        Really deletes the run and its report. If the run is still `running` it is
        stopped first, then the record is removed — so this both aborts a run and
        cleans up finished history in one step. Unlike the cancel action, nothing
        survives: a subsequent `GET` is `404`. Either verb frees a concurrency slot;
        cancel keeps the report, delete removes it.
      tags: [system]
      responses:
        "204":
          description: The run was stopped (if running) and its report deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /system/verification/{reportId}/cancel:
    parameters:
      - name: reportId
        in: path
        required: true
        description: The verification report id.
        schema: { type: string }
    post:
      operationId: cancelVerification
      summary: Cancel a running verification run
      description: |
        Stops a `running` run but **keeps** its report: the record stays in history
        with `status` `stopped` and whatever partial findings it had gathered, and the
        concurrency slot is freed. Use this to abort a run you want to keep a record
        of; use `DELETE` to remove it entirely. Idempotent — cancelling a run that has
        already finished or stopped returns its current report unchanged.
      tags: [system]
      responses:
        "200":
          description: The run's report after the cancel (status `stopped` if it was running).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "`Authorization: Bearer <jwt>` → the JWT auth adapter."
    apikey:
      type: apiKey
      in: header
      name: X-API-Key
      description: "`X-API-Key: <key>` → the API-key auth adapter."
    macaroon:
      type: http
      scheme: macaroon
      description: "`Authorization: Macaroon <token>` → the macaroon auth adapter."
  parameters:
    BranchName:
      name: branch
      in: path
      required: true
      description: The branch name (`$` is reserved and illegal in ordinary names).
      schema: { type: string }
    ClaimId:
      name: id
      in: path
      required: true
      description: The content-addressed claim id.
      schema: { type: string }
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication is required or failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        Access was denied by core-access. The body carries no subject id or
        onboarding hint.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        The branch, claim, or content is unknown — or lies outside the named
        branch's closure. The two are indistinguishable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    BranchHead:
      type: object
      required: [head]
      properties:
        head:
          type: string
          description: The content-addressed head claim id.
...
    VerificationConfig:
      type: object
      required: [closure]
      description: |
        Parameters for a verification run — the same shape whether declared in the
        stack config (scheduled) or posted ad-hoc. Depths are those of
        core-verification.
      properties:
        closure:
          type: string
          description: Root of the closure to verify — a branch name or a claim id.
        layer:
          type: string
          description: |
            Storage layer (by name) to read directly. Omit to read the composed
            read-through view — which may mask the loss of an object on a deeper
            layer, so name a layer to verify it without blind spots.
        depth:
          type: string
          enum: [completeness, record-correctness, full-content]
          description: |
            completeness (a `has` sweep), record-correctness (recanonicalise and
            recheck the id-chain and signatures), or full-content (also re-hash blobs).
        contentThreshold:
          type: integer
          description: |
            For full-content, the max content size in bytes to re-read and re-hash
            per claim; larger content is skipped this run. Omit to verify all content.
    VerificationReport:
      type: object
      required: [id, config, head, status, startedAt, ok]
      description: A point-in-time record of a verification run; embeds the config that produced it.
      properties:
        id:
          type: string
        config:
          $ref: "#/components/schemas/VerificationConfig"
        head:
          type: string
          description: |
            The head id the run verified — the closure it pinned at start. When
            `config.closure` names a branch, this is the branch's head at start time;
            the report stays fixed to it even as the branch head moves on.
        status:
          type: string
          enum: [running, complete, stopped, error]
          description: |
            running (in progress), complete (finished on its own), stopped (cancelled
            by an operator via the cancel action — partial findings kept), or error
            (the run itself failed).
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        claimsChecked:
          type: integer
          description: Claims checked so far; advances while `status` is `running`.
        bytesRead:
          type: integer
          description: Content bytes re-read so far; advances while `status` is `running`.
        ok:
          type: boolean
          description: True when no failures were found in this run.
        failures:
          type: array
          description: |
            The failures found, accumulating while `status` is `running`. Empty (with
            `ok: true`) for an intact closure; a healthy archive verifies with none.
          items: { $ref: "#/components/schemas/VerificationFailure" }
    VerificationReportList:
      type: object
      required: [reports]
      properties:
        reports:
          type: array
          items: { $ref: "#/components/schemas/VerificationReport" }
    VerificationFailure:
      type: object
      required: [id, mode, layer]
      properties:
        id:
          type: string
          description: The claim or object id that failed.
        mode:
          type: string
          enum: [corrupt-bytes, invalid-content]
          description: |
            corrupt-bytes — stored bytes don't match their hash (storage rot/loss;
            self-heals via read-through if a deeper layer is intact).
            invalid-content — the claim itself doesn't validate (e.g. bad signature);
            unrepairable.
        layer:
          type: string
          description: The layer where the failure was observed.
        detail:
          type: string
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: |
            Machine-readable failure category, stable across releases — one of
            unauthenticated, forbidden, not_found, conflict, busy, invalid,
            unimplemented, internal. Clients branch on this, not on the message.
        error:
          type: string
          description: Human-readable message. Carries no subject id, even on 403.